requirements.txt
streamlit
Pillow
requests
google-cloud-genai
